requirements.txt
numpy
Pillow
Flask
pycuda; platform_system != 'Windows'
